Se agrega funcionalidad para vincular certificado, se modifican las alertas en catalogos y autenticación

+++ src/stores/catalogoStore.js
@@ -29,53 +29,11 @@ export const useCatalogoStore = defineStore('catalogoStore',() => {
         catPuesto: [],
         catSexo: []
     });
-    const loadingArea = ref(false);
     const loadingCat = ref(false);
-    const getCatArea = async(token) => {
-        const urlCat = baseUrl + "/api/get-catalogo/areas";
-        try {
-            //const { data } = await axios.get(urlCat, {headers:{"Authorization": `Bearer ${token}`}});
-            const { data } = await axios.get(urlCat, getAuthorizationHeadersForLaravel(token));
-            catState.value.catArea = data;
-        } catch (error) {
-            console.log(error);
-        }finally{
-           //loadingArea.value = false;
-        }
-    }
-    const getCatPuesto = async(token) => {
-        const urlCat = baseUrl + "/api/get-catalogo/puestos";
-        try {
-            //const { data } = await axios.get(urlCat, {headers:{"Authorization": `Bearer ${token}`}});
-            const { data } = await axios.get(urlCat, getAuthorizationHeadersForLaravel(token));
-            // catState.value.catPuesto = data;
-            return data;
-        } catch (error) {
-            console.log(error);
-        }finally{
-           //loadingArea.value = false;
-        }
-    }
-    const getCatSexo = async(token) => {
-        const urlCat = baseUrl + "/api/get-catalogo/sexo";        
-        try {
-            //const { data } = await axios.get(urlCat, {headers:{"Authorization": `Bearer ${token}`}});
-            const { data } = await axios.get(urlCat, getAuthorizationHeadersForLaravel(token));
-            
-            return data;
-        } catch (error) {
-            console.log(error);
-        }finally{
-           //loadingArea.value = false;
-        }
-    }
 
     const getCatalogo = async(url, token) => {
         const urlCat = baseUrl + url;
         try {
-            //local
-            //const { data } = await axios.get(urlCat, {headers:{"Authorization": `Bearer ${token}`}});
-            //servidor
             const { data } = await axios.get(urlCat, getAuthorizationHeadersForLaravel(token));
             return data;
         } catch (error) {
@@ -88,17 +46,12 @@ export const useCatalogoStore = defineStore('catalogoStore',() => {
         const urlSaveCat = baseUrl + urlCat;
         
         try {
-          //   const axiosInstance = axios.create({
-          //      "Access-Control-Allow-Origin": "*",
-          //  });
-           //local
-           //await axios.post(urlSaveCat, data, {headers:{"Authorization": `Bearer ${token}`}}).then((response) => {
-            //servidor
             await axios.post(urlSaveCat, data, getAuthorizationHeadersForLaravel(token)).then((response) => {
               //console.log("Resp",response)
-                 if (confirm(response.data.message)) {
-                      window.location.reload();
-                    }
+              showMessage('\u00A1Guardado!');
+              setTimeout(()=>{
+                window.location.reload();
+              }, 800);
                 });
             
            } catch (error) {
@@ -108,13 +61,11 @@ export const useCatalogoStore = defineStore('catalogoStore',() => {
     const editCatalogo = async(urlCat, data, token) => {
         const urlEditCat = baseUrl + urlCat;
         try {
-            //local
-                //await axios.put(urlEditCat, data, {headers:{"Authorization": `Bearer ${token}`}}).then((response) => {
-            //servidor
                 await axios.put(urlEditCat, data, getAuthorizationHeadersForLaravel(token)).then((response) => {
-                 if (confirm(response.data.message)) {
-                      window.location.reload();
-                   }
+                  showMessage('\u00A1Editado!');
+                  setTimeout(()=>{
+                    window.location.reload();
+                  }, 800);
                 });
             
            } catch (error) {
@@ -124,14 +75,24 @@ export const useCatalogoStore = defineStore('catalogoStore',() => {
     const deleteCatalogo = async(urlCat, token) => {
         const urlDelCat = baseUrl + urlCat;
         try {
-            //local
-            //await axios.delete(urlDelCat, {headers:{"Authorization": `Bearer ${token}`}}).then((response) => {
-            //servidor
-            await axios.delete(urlDelCat, getAuthorizationHeadersForLaravel(token)).then((response) => {
-                  if (confirm(response.data.message)) {
-                       window.location.reload();
-                    }
-                 });
+          new window.Swal({
+              icon: 'warning',
+              title: '\u00BFDesea eliminar este elemento?',
+              text: "Esta opci\u00F3n sera eliminada definitivamente",
+              showCancelButton: true,
+              confirmButtonText: 'Eliminar',
+              padding: '2em',
+          }).then(async(result) => {
+              if (result.value) {
+                await axios.delete(urlDelCat, getAuthorizationHeadersForLaravel(token)).then((response) => {
+                  //console.log(response.data)
+                  showMessage('\u00A1Eliminado!');
+                  setTimeout(()=>{
+                    window.location.reload();
+                  }, 800);
+                });
+              }
+          });
             
            } catch (error) {
              console.log(error)
@@ -139,9 +100,6 @@ export const useCatalogoStore = defineStore('catalogoStore',() => {
     }
     const getNuevoDocumento = async(urlNewDoc, token) => {
         try {
-            //local
-             //const { data } = await axios.get(urlNewDoc, {headers:{"Authorization": `Bearer ${token}`}});
-             //servidor
              const { data } = await axios.get(urlNewDoc, getAuthorizationHeadersForLaravel(token));
              return data;    
            } catch (error) {
@@ -166,11 +124,22 @@ export const useCatalogoStore = defineStore('catalogoStore',() => {
          }
   }
 
+  //Alerts
+  const showMessage = (msg = '', type = 'success') => {
+    const toast = window.Swal.mixin({
+        toast: true,
+        position: 'top',
+        showConfirmButton: false,
+        timer: 3000,
+    });
+    toast.fire({
+        icon: type,
+        title: msg,
+        padding: '10px 20px',
+    });
+};
     return { 
         catState, 
-        //getCatArea, 
-        //getCatPuesto, 
-        //getCatSexo, 
         getCatalogo, 
         saveCatalogo, 
         editCatalogo,
